// backend/tests/__mocks__/@prisma/client.d.ts
declare const mockPrisma: {
    user: {
        create: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findMany: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        update: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
        deleteMany: jest.Mock<any, any, any>;
        count: jest.Mock<any, any, any>;
    };
    order: {
        create: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findMany: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        update: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
        deleteMany: jest.Mock<any, any, any>;
        count: jest.Mock<any, any, any>;
    };
    cart: {
        create: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        update: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
        deleteMany: jest.Mock<any, any, any>;
    };
    article: {
        create: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findMany: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        update: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
        deleteMany: jest.Mock<any, any, any>;
        count: jest.Mock<any, any, any>;
    };
    comment: {
        create: jest.Mock<any, any, any>;
        findMany: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
        deleteMany: jest.Mock<any, any, any>;
    };
    like: {
        create: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
        deleteMany: jest.Mock<any, any, any>;
    };
    service: {
        create: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findMany: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        update: jest.Mock<any, any, any>;
        delete: jest.Mock<any, any, any>;
    };
    setting: {
        findMany: jest.Mock<any, any, any>;
        findUnique: jest.Mock<any, any, any>;
        findFirst: jest.Mock<any, any, any>;
        upsert: jest.Mock<any, any, any>;
    };
    $transaction: jest.Mock<any, any, any>;
    $disconnect: jest.Mock<any, any, any>;
};